# Faster identifier tokenizing