--- a/packages/babel-parser/src/tokenizer/context.js
+++ b/packages/babel-parser/src/tokenizer/context.js
@@ -76,10 +76,6 @@ tt.name.updateContext = function (prevType) {
     }
   }
   this.state.exprAllowed = allowed;
-
-  if (this.state.isIterator) {
-    this.state.isIterator = false;
-  }
 };
 
 tt.braceL.updateContext = function (prevType) {